my first Interview
My first interview is in my school,the interviewer give me a question,how to reverse a string? I can use string,or character array,I chose the character,but I think my code is not good,so when I came back to my dormitory,and I do it again,and decide to save it. I just a beginner to learn C++, so ,maybe my code look not so good.